Create gen_excel_xor_obfuscation_velvetsweatshop.yar
This detects malware using an ancient "encryption" algo (XOR and the default VelvetSweatshop password)--it was removed from the UX as an option and last showed up in Office 2003, but current Office versions will open documents encrypted with it. There was a big malware campaign that used this and this was the only yara rule that detected it.  (see more here: https://twitter.com/JohnLaTwC/status/1314602421977452544) .  More matching samples found here (https://twitter.com/JohnLaTwC/status/1314629735905857536).  This rule was created by @BouncyHat and I am submitting it on his behalf to get it into yara repo.