Proposition of change based on follina research.
Hi,

I've observed that you can put wahtever after the '!' hence the changes (https://twitter.com/bodziurity/status/1531618358071795712). Also the .htm extension works not just .html. Updated rule won't trigger only for follina though, it triggers on MSHTML CVE-2021-40444 as well, if you'll accept this proposition I'll leave the creating this as a different rule or leaving at it is to you.

PS. I did not play with rtf version of the exploit. 
PS2. My box has Office 2016 and not the latest version, might be good to check on other versions too :)

Regards,
Wojciech Cieślak